+++ src/main/java/team/tab/daixu/service/impl/ArticleLabelServiceImpl.java
@@ -16,7 +16,7 @@ public class ArticleLabelServiceImpl implements ArticleLabelService {
     private ArticleLabelDao articleLabelDaoImpl;
 
     @Override
-    public Boolean actSave(int article_id, String mix_article_label) {
+    public Boolean actSave(long article_id, String mix_article_label) {
 //        String[] array_label = mix_article_label.split(",");
 //
 //        Boolean result_save = articleLabelDaoImpl.act_save(article_id,array_label);
